feat: add logs service
