+++ back/apps/barramento/package.json
@@ -13,8 +13,8 @@
   "license": "ISC",
   "dependencies": {
     "axios": "1.7.2",
-    "better-sqlite3": "10.0.0",
     "cors": "2.8.5",
+    "dotenv": "16.4.5",
     "express": "4.19.2",
     "morgan": "1.10.0",
     "nodemon": "3.1.0"
