--- back/apps/bff/.env.sample
@@ -1,1 +1,0 @@
-OPENAI_API_KEY=sk-some-key
