+++ back/apps/logs/package.json
@@ -16,9 +16,7 @@
     "cors": "2.8.5",
     "dotenv": "16.4.5",
     "express": "4.19.2",
-    "helmet": "7.1.0",
-    "morgan": "1.10.0",
-    "openai": "4.38.5"
+    "morgan": "1.10.0"
   },
   "devDependencies": {
     "nodemon": "3.1.0"
