+++ back/apps/logs/src/db/logs.manager.js
