+++ back/apps/logs/src/server.js
@@ -1,18 +1,21 @@
-import express from 'express';
-import { router } from './router.js';
-import helmet from 'helmet';
-import morgan from 'morgan';
 import cors from 'cors';
+import dotenv from 'dotenv';
+import express from 'express';
+import morgan from 'morgan';
+import { router } from './router.js';
+
+dotenv.config();
 
 const app = express();
 
 app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
-app.use(helmet());
 app.use(cors());
 app.use(morgan(':method :url :status :response-time'));
 app.use(router);
 
-app.listen(3000, () => {
-  console.log('Server is running at http://localhost:3000');
+const port = process.env.PORT || 4000;
+
+app.listen(port, () => {
+  console.log('Logs listening at http://localhost:4000');
 });
