+++ back/apps/message/package.json
@@ -12,13 +12,13 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "axios": "^1.7.2",
+    "axios": "1.7.2",
     "better-sqlite3": "10.0.0",
     "cors": "2.8.5",
-    "dotenv": "^16.4.5",
+    "dotenv": "16.4.5",
     "express": "4.19.2",
     "morgan": "1.10.0",
     "nodemon": "3.1.0",
-    "openai": "^4.47.1"
+    "openai": "4.47.1"
   }
 }
