+++ back/apps/message/src/server.js
@@ -1,7 +1,10 @@
 import cors from 'cors';
+import dotenv from 'dotenv';
 import express from 'express';
 import morgan from 'morgan';
 import { router } from './router.js';
+
+dotenv.config();
 
 const app = express();
